Add files via upload
An iris machine learning model is a tool that classifies different types of iris flowers based on their features, such as petal and sepal length and width. It learns from a dataset of known iris species (setosa, versicolor, and virginica) and uses this knowledge to accurately identify the species of new iris flowers.